+++ Source/Boilerplate.Templates/Content/ApiTemplate/Startup.cs
@@ -1,5 +1,8 @@
 ﻿namespace ApiTemplate
 {
+#if (Versioning)
+    using System.Linq;
+#endif
 #if (CORS)
     using ApiTemplate.Constants;
 #endif
@@ -8,6 +11,9 @@
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Mvc;
+#if (Versioning)
+    using Microsoft.AspNetCore.Mvc.ApiExplorer;
+#endif
     using Microsoft.AspNetCore.Mvc.Infrastructure;
     using Microsoft.AspNetCore.Mvc.Routing;
 #if (HttpsEverywhere)
@@ -108,9 +114,7 @@
         /// called by the ASP.NET runtime. See
         /// http://blogs.msdn.com/b/webdev/archive/2014/06/17/dependency-injection-in-asp-net-vnext.aspx
         /// </summary>
-        /// <param name="services">The services collection or IoC container.</param>
-        public void ConfigureServices(IServiceCollection services)
-        {
+        public void ConfigureServices(IServiceCollection services) =>
             services
 #if (ApplicationInsights)
                 // Add Azure Application Insights data collection services to the services container.
@@ -135,7 +139,7 @@
 #if (Versioning)
                 .AddCustomVersioning()
 #endif
-                .AddCustomMvc(this.configuration, this.hostingEnvironment)
+                .AddMvcCore()
                 .AddApiExplorer()
                 .AddAuthorization()
                 .AddFormatterMappings()
@@ -152,18 +156,19 @@
                 // Adds the XML input and output formatter using the XmlSerializer.
                 .AddXmlSerializerFormatters()
 #endif
+#if (Swagger && Versioning)
+                .AddVersionedApiExplorer()
+#endif
+                .AddCustomMvcOptions(this.configuration, this.hostingEnvironment)
                 .Services
                 .AddCommands()
                 .AddRepositories()
                 .AddTranslators();
-        }
 
         /// <summary>
         /// Configures the application and HTTP request pipeline. Configure is called after ConfigureServices is
         /// called by the ASP.NET runtime.
         /// </summary>
-        /// <param name="application">The application.</param>
-        /// <param name="loggerfactory">The logger factory.</param>
         public void Configure(IApplicationBuilder application, ILoggerFactory loggerfactory)
         {
             // Configure application logging. See http://docs.asp.net/en/latest/fundamentals/logging.html
@@ -178,8 +183,6 @@
                         .AddDebug());
 
             application
-                // Removes the Server HTTP header from the HTTP response for marginally better security and performance.
-                .UseNoServerHttpHeader()
 #if (HttpsEverywhere)
                 // Require HTTPS to be used across the whole site. Also set a custom port to use for SSL in
                 // Development. The port number to use is taken from the launchSettings.json file which Visual
@@ -204,15 +207,25 @@
                 .UsePublicKeyPinsHttpHeader()
 #endif
 #endif
-                // Add MVC to the request pipeline.
 #if (Swagger)
                 .UseMvc()
-                // Add Swagger to the request pipeline.
                 .UseSwagger()
-                .UseSwaggerUi(
+                .UseSwaggerUI(
                     options =>
                     {
+#if (Versioning)
+                        var provider = application.ApplicationServices.GetService<IApiVersionDescriptionProvider>();
+                        foreach (var apiVersionDescription in provider
+                            .ApiVersionDescriptions
+                            .OrderByDescending(x => x.ApiVersion))
+                        {
+                            options.SwaggerEndpoint(
+                                $"/swagger/{apiVersionDescription.GroupName}/swagger.json",
+                                $"Version {apiVersionDescription.ApiVersion}");
+                        }
+#else
                         options.SwaggerEndpoint("/swagger/v1/swagger.json", "Version 1");
+#endif
                     });
 #else
                 .UseMvc();