+++ RELEASE NOTES.md
@@ -1,3 +1,12 @@
+# Boilerplate.Templates 1.5.0
+- Added API Versioning support for Swagger.
+
+# Boilerplate.Templates 1.4.0
+- Added OPTIONS endpoint.
+- Added HEAD endpoints to all GET endpoints.
+- Update Swashbuckle.AspNetCore to 1.0.0-rc3
+- Use AddMvcOptions instead the delegate in AddMvcCore.
+
 # Boilerplate.Templates 1.3.0
 - Added Versioning feature.
 - Add Link HTTP Header to GET /cars to show next, previous, first, last page URL.